refactor(Component): :recycle: Polish component references
