+++ src/config/routes.ts
@@ -1,5 +1,5 @@
+import { ComponentsPaths } from '../pages/Components/Components.constants';
 import { ProcessesPaths } from '../pages/Processes/Processes.constants';
-import { ComponentsPaths } from '../pages/ProcessGroups/Components.constants';
 import { ServicesPaths } from '../pages/Services/Services.constants';
 import { SitesPaths } from '../pages/Sites/Sites.constants';
 import { TopologyPaths } from '../pages/Topology/Topology.constants';
