+++ src/pages/Components/hooks/useComponentPairsData.ts
