+++ src/pages/Components/views/Component.tsx
