+++ src/pages/Topology/components/NodeDetails.tsx
@@ -28,8 +28,8 @@ import { ellipsisInTheMiddle } from '../../../core/utils/EllipsisInTheMiddle';
 import { formatByteRate, formatBytes } from '../../../core/utils/formatBytes';
 import { ProcessResponse } from '../../../types/REST.interfaces';
 import { TopologyMetrics } from '../../../types/Topology.interfaces';
+import { ComponentRoutesPaths } from '../../Components/Components.enum';
 import { ProcessesLabels, ProcessesRoutesPaths } from '../../Processes/Processes.enum';
-import { ComponentRoutesPaths } from '../../ProcessGroups/Components.enum';
 import { MetricsLabels } from '../../shared/Metrics/Metrics.enum';
 import { SitesRoutesPaths } from '../../Sites/Sites.enum';
 import { TopologyLabels } from '../Topology.enum';
