+++ src/pages/Topology/views/Topology.tsx
@@ -6,8 +6,8 @@ import { useSearchParams } from 'react-router-dom';
 import { getTestsIds } from '../../../config/testIds';
 import useUpdateQueryStringValueWithoutNavigation from '../../../hooks/useUpdateQueryStringValueWithoutNavigation';
 import MainContainer from '../../../layout/MainContainer';
+import { ComponentRoutesPaths } from '../../Components/Components.enum';
 import { ProcessesRoutesPaths } from '../../Processes/Processes.enum';
-import { ComponentRoutesPaths } from '../../ProcessGroups/Components.enum';
 import { SitesRoutesPaths } from '../../Sites/Sites.enum';
 import TopologyComponent from '../components/TopologyComponent';
 import TopologyProcesses from '../components/TopologyProcesses';
