+++ src/routes.tsx
@@ -1,8 +1,8 @@
 import { Navigate, RouteObject } from 'react-router-dom';
 
 import { DEFAULT_ROUTE } from './config/routes';
+import { componentsRoutes } from './pages/Components/routes';
 import { processesRoutes } from './pages/Processes/routes';
-import { processGroupsRoutes } from './pages/ProcessGroups/routes';
 import { servicesRoutes } from './pages/Services/routes';
 import { errorsRoutes } from './pages/shared/Errors/routes';
 import { siteRoutes } from './pages/Sites/routes';
@@ -12,7 +12,7 @@ export const routes: RouteObject[] = [
   { index: true, element: <Navigate to={DEFAULT_ROUTE} replace={true} /> },
   ...siteRoutes,
   ...servicesRoutes,
-  ...processGroupsRoutes,
+  ...componentsRoutes,
   ...processesRoutes,
   ...errorsRoutes,
   ...topologyRoutes
